requirements.txt
pytest
tensorflow-gpu==2.2
librosa~=0.8.0
matplotlib~=3.3.3
numpy==1.17.2
scipy==1.4.1
sklearn~=0.0

scikit-learn~=0.23.2
SoundFile~=0.10.3.post1